requirements.txt
# AWS Strands Agents SDK - Core packages
strands-agents>=1.9.0
strands-tools>=0.2.8

# AWS Services
boto3>=1.34.0
botocore>=1.34.0

# Academic Paper APIs
arxiv>=2.1.0
requests>=2.31.0

# PDF Processing
PyMuPDF>=1.23.0

# Text Processing
nltk>=3.8.1
beautifulsoup4>=4.12.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Environment and Configuration
python-dotenv>=1.0.0

# JSON and Data Handling
simplejson>=3.19.0

# HTTP and Networking
urllib3>=2.0.0
certifi>=2023.7.22
charset-normalizer>=3.2.0
idna>=3.4

# AWS Lambda Dependencies
python-dateutil>=2.8.2
jmespath>=1.0.1
s3transfer>=0.7.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0

# Additional utilities
six>=1.16.0
setuptools>=68.0.0
wheel>=0.41.0
